# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: BlueStrongMountain API
  version: 1.0.0
  description: Swagger spec for BoardController and ProblemFilterController
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/groups:
    post:
      tags: [Group]
      summary: Create a new group
      parameters:
        - in: query
          name: requesterId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreateRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'

    get:
      tags: [Group]
      summary: Get my groups or search my groups
      parameters:
        - in: query
          name: requesterId
          required: true
          schema:
            type: integer
        - in: query
          name: name
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupSummaryResponse'
  /api/v1/groups/{groupId}:
    patch:
      tags: [Group]
      summary: Update group information
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
        - in: query
          name: requesterId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupUpdateRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
    delete:
      tags:
        - Group
      summary: Delete group
      description: |
        그룹을 삭제합니다.
        - 요청자는 해당 그룹의 OWNER여야 합니다.
        - 그룹에 속한 모든 유저 및 관련 진행 정보가 함께 삭제됩니다.
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
            format: int64
          description: 삭제할 그룹 ID
        - in: query
          name: requesterId
          required: true
          schema:
            type: integer
            format: int64
          description: 그룹 삭제를 요청한 사용자 ID (OWNER)
      responses:
        '200':
          description: Group deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
              example:
                success: true
                message: "OK"
        '403':
          description: Forbidden - requester is not group owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
              example:
                success: false
                message: "Not authorized to delete this group"
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
              example:
                success: false
                message: "Group not found"
  /api/v1/groups/{groupId}/detail:
    get:
      tags: [Group]
      summary: "그룹 상세 조회"
      description: "요청자가 속한 그룹의 상세 정보를 조회합니다."
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
          description: "조회할 그룹 ID"

        - name: requesterId
          in: query
          required: true
          schema:
            type: integer
          description: "요청자의 사용자 ID"

      responses:
        '200':
          description: "그룹 상세 정보 반환"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDetailDto'

        '404':
          description: "그룹을 찾을 수 없음"
        '403':
          description: "그룹 접근 권한 없음"
  /api/v1/groups/{groupId}/users:
    get:
      tags:
        - Group
      summary: Get all users in a group
      description: 그룹에 속한 모든 유저 목록을 조회한다.
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
          description: 그룹 ID
        - in: query
          name: requesterId
          required: true
          schema:
            type: integer
          description: 요청자 사용자 ID
      responses:
        '200':
          description: Group users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupUserDto'
        '403':
          description: Permission denied
        '404':
          description: User or group not found
  /api/v1/groups/{groupId}/owner:
    patch:
      tags: [Group]
      summary: Transfer group ownership
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
        - in: query
          name: requesterId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnerChangeRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
  /api/v1/groups/{groupId}/managers:
    post:
      tags: [Group]
      summary: Add managers to group (bulk)
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
        - in: query
          name: requesterId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIdsRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'

    delete:
      tags: [Group]
      summary: Remove managers (bulk)
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
        - in: query
          name: requesterId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManagerIdsRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
  /api/v1/groups/{groupId}/members:
    post:
      tags: [Group]
      summary: Add members to group (bulk)
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
        - in: query
          name: requesterId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIdsRequest'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'

    delete:
      tags: [Group]
      summary: Remove members (bulk)
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
        - in: query
          name: requesterId
          required: true
          schema:
            type: integer
        - in: query
          name: userIds
          required: true
          schema:
            type: array
            items:
              type: integer
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
  /api/v1/groups/{groupId}/members/me:
    delete:
      tags: [Group]
      summary: Leave group (self)
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            type: integer
        - in: query
          name: requesterId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'


  ########################################################
  # BoardController APIs
  ########################################################
...
components:
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: OK
      ############################################################
    # Auth DTOs
    ############################################################
...
    GroupCreateRequest:
      type: object
      properties:
        title:
          type: string
        managerIds:
          type: array
          items:
            type: integer
        memberIds:
          type: array
          items:
            type: integer
        visibility:
          type: string
        description:
          type: string
    GroupUpdateRequest:
      type: object
      properties:
        title:
          type: string
        managerIds:
          type: array
          items:
            type: integer
        memberIds:
          type: array
          items:
            type: integer
        visibility:
          type: string
        description:
          type: string
    GroupSummaryResponse:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        ownerId: { type: integer }
        visibility: { type: string }
        memberCount: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
    GroupDetailDto:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        ownerId:
          type: integer
        managerIds:
          type: array
          items:
            type: integer
        memberIds:
          type: array
          items:
            type: integer
        visibility:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    BasicResponse:
      type: object
      properties:
        success:
          type: boolean
    UserIdsRequest:
      type: object
      properties:
        userIds:
          type: array
          items:
            type: integer
    ManagerIdsRequest:
      type: object
      properties:
        managerIds:
          type: array
          items:
            type: integer
    OwnerChangeRequest:
      type: object
      properties:
        newOwnerId:
          type: integer
    GroupUserDto:
      type: object
      properties:
        userId:
          type: integer
          example: 1
        username:
          type: string
          example: fermat001
        role:
          type: string
          enum:
            - OWNER
            - MANAGER
            - MEMBER
          example: MEMBER

      ############################################################
    # Board DTOs
    ############################################################
...